Create BaseAddCardFragment base class
- Create `BaseAddCardFragment` abstract class
- Move logic from `PaymentSheetAddCardFragment` to `BaseAddCardFragment`
- Create new `PaymentSheetAddCardFragment` and
  `PaymentOptionsAddCardFragment` that extend `BaseAddCardFragment`
- Use `PaymentOptionsAddCardFragment` in `PaymentOptionsActivity`
